+++ packages/interop-backoffice-ui/src/pages/summary/table/columns.tsx
@@ -1,33 +1,7 @@
-import type { ColumnDef, HeaderContext } from '@tanstack/react-table'
-import { ArrowDownIcon, ArrowUpDownIcon, ArrowUpIcon } from 'lucide-react'
+import type { ColumnDef } from '@tanstack/react-table'
 import { Link } from 'react-router-dom'
-import { Button } from '~/components/core/Button'
+import { SortableHeader } from '~/components/table/SortableHeader'
 import type { SummaryEventRow } from './types'
-
-function SortableHeader(
-  props: HeaderContext<SummaryEventRow, unknown> & { label: string },
-) {
-  const { column, label } = props
-  const sortState = column.getIsSorted()
-
-  return (
-    <Button
-      variant="ghost"
-      size="sm"
-      className="-ml-3"
-      onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-    >
-      {label}
-      {sortState === 'asc' ? (
-        <ArrowUpIcon className="text-primary" />
-      ) : sortState === 'desc' ? (
-        <ArrowDownIcon className="text-primary" />
-      ) : (
-        <ArrowUpDownIcon className="opacity-50" />
-      )}
-    </Button>
-  )
-}
 
 export const eventsColumns: ColumnDef<SummaryEventRow>[] = [
   {
