+++ packages/interop-backoffice-ui/src/pages/summary/table/transfers/pairs-columns.tsx
@@ -1,33 +1,7 @@
-import type { ColumnDef, HeaderContext } from '@tanstack/react-table'
-import { ArrowDownIcon, ArrowUpDownIcon, ArrowUpIcon } from 'lucide-react'
-import { Button } from '~/components/core/Button'
+import type { ColumnDef } from '@tanstack/react-table'
+import { SortableHeader } from '~/components/table/SortableHeader'
 import type { SummaryTransferPairRow } from '../types'
 import { formatDollars, formatDuration } from './utils'
-
-function SortableHeader(
-  props: HeaderContext<SummaryTransferPairRow, unknown> & { label: string },
-) {
-  const { column, label } = props
-  const sortState = column.getIsSorted()
-
-  return (
-    <Button
-      variant="ghost"
-      size="sm"
-      className="-ml-3"
-      onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-    >
-      {label}
-      {sortState === 'asc' ? (
-        <ArrowUpIcon className="text-primary" />
-      ) : sortState === 'desc' ? (
-        <ArrowDownIcon className="text-primary" />
-      ) : (
-        <ArrowUpDownIcon className="opacity-50" />
-      )}
-    </Button>
-  )
-}
 
 export const transferPairsColumns: ColumnDef<SummaryTransferPairRow>[] = [
   {
